+++ changelog.md
@@ -5,9 +5,10 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Replace Dependabot by Renovate for dependency updates. Closes #1728.
+- Use Renovate to sync `uv.lock` files. Remove `sync-uv-lock` job.
 - Add `toml_files` field to `gha-utils metadata` output.
 - Replace `julb/action-manage-label` by `labelmaker` for label management. Closes #1914.
-- Fix Dependabot labels synchronization.
 - Add Rust-related entries to `.gitignore` file.
 
 ## [4.25.5 (2026-01-09)](https://github.com/kdeldycke/workflows/compare/v4.25.4...v4.25.5)